Drop Plone 3 and 4, add Plone 5 and 6.
This is only initial support.
The Plone 6 instance I am trying this in has other packages which do not start yet.
But at least all ImportErrors seem gone, and we no longer use zope.formlib.
May need a different widget.

Current stable version is for Plone 4 and does not work on 5.
I want it to work on Plone 6, and will use `collective.exportimport` to directly go there.
But in order to have a more standard migration path, a new version should support Plone 5.2 on Python 2 as well.
This PR does that.